requirements.txt
setuptools>=40.0.0
streamlit==1.25.0
pillow==10.0.0
numpy==1.21.0
pandas==1.3.3